pendant/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==11.0.3
numpy==1.24.3
pyserial==3.5
python-multipart==0.0.6
pydantic==2.4.2
